Move visitors to separate packages and add missing potentialHazard visitors
