+++ core/src/main/java/com/graphhopper/reader/osgb/dpn/potentialHazards/Spoil.java
@@ -1,6 +1,7 @@
-package com.graphhopper.reader.osgb.dpn;
+package com.graphhopper.reader.osgb.dpn.potentialHazards;
 
 import com.graphhopper.reader.Way;
+import com.graphhopper.reader.osgb.dpn.AbstractOsDpnOsmAttibuteMappingVisitor;
 
 /**
  * Created by sadam on 13/02/15.
