+++ core/src/test/java/com/graphhopper/reader/osgb/dpn/InlandWaterTest.java
@@ -9,6 +9,7 @@ import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 
 import com.graphhopper.reader.Way;
+import com.graphhopper.reader.osgb.dpn.potentialHazards.InlandWater;
 
 public class InlandWaterTest {
 	 static OsDpnOsmAttributeMappingVisitor visitor;
