+++ core/src/test/java/com/graphhopper/reader/osgb/dpn/NoneTest.java
@@ -1,6 +1,8 @@
 package com.graphhopper.reader.osgb.dpn;
 
 import com.graphhopper.reader.Way;
+import com.graphhopper.reader.osgb.dpn.rightOfWay.None;
+
 import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
