+++ core/src/test/java/com/graphhopper/reader/osgb/dpn/QuarryOrPitTest.java
@@ -10,6 +10,7 @@ import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 
 import com.graphhopper.reader.Way;
+import com.graphhopper.reader.osgb.dpn.potentialHazards.QuarryOrPit;
 
 public class QuarryOrPitTest {
 	 static OsDpnOsmAttributeMappingVisitor visitor;
