+++ core/src/test/java/com/graphhopper/reader/osgb/dpn/SandTest.java
@@ -9,6 +9,7 @@ import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 
 import com.graphhopper.reader.Way;
+import com.graphhopper.reader.osgb.dpn.potentialHazards.Sand;
 
 public class SandTest {
 	 static OsDpnOsmAttributeMappingVisitor visitor;
