+++ core/src/test/java/com/graphhopper/reader/osgb/dpn/TidalWaterTest.java
@@ -9,6 +9,7 @@ import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 
 import com.graphhopper.reader.Way;
+import com.graphhopper.reader.osgb.dpn.potentialHazards.TidalWater;
 
 public class TidalWaterTest {
 	 static OsDpnOsmAttributeMappingVisitor visitor;
